pyproject.toml
[project]
name = "personal_notion_reading_note"
version = "0.1.0"
description = "A tool to generate AI reading notes and push to notion."
dependencies = [
    "google-generativeai",
    "requests",
    "python-dotenv",
    "PyYAML",
    "pypdf",
    "dspy-ai",
    "pydantic",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"